+++ tofhir-server/src/main/scala/io/tofhir/server/service/JobService.scala
@@ -1,17 +1,9 @@
 package io.tofhir.server.service
 
 import com.typesafe.scalalogging.LazyLogging
-import io.tofhir.engine.model.{FhirMappingJob, FhirMappingResult, FhirMappingTask}
-import io.tofhir.server.config.SparkConfig
-import io.tofhir.server.model.{TestResourceCreationRequest, ResourceNotFound}
+import io.tofhir.engine.model.FhirMappingJob
 import io.tofhir.server.service.job.IJobRepository
-import org.apache.spark.sql.catalyst.encoders.RowEncoder
-import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
-import org.apache.spark.sql.{Encoders, Row}
-import org.json4s.JsonAST.JValue
-import org.json4s.jackson.JsonMethods
 
-import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
 class JobService(jobRepository: IJobRepository) extends LazyLogging {
@@ -64,131 +56,5 @@ class JobService(jobRepository: IJobRepository) extends LazyLogging {
    */
   def deleteJob(projectId: String, id: String): Future[Unit] = {
     jobRepository.deleteJob(projectId, id)
-  }
-
-  /**
-   * Run the job for the given mapping tasks
-   * @param projectId project id the job belongs to
-   * @param jobId job id
-   * @param mappingUrls the mapping tasks to be executed
-   * @return
-   */
-  def runJob(projectId: String, jobId: String,mappingUrls: Option[Seq[String]]=None): Future[Future[Unit]] = {
-    jobRepository.runJob(projectId, jobId,mappingUrls)
-  }
-
-  /**
-   * Tests the given mapping task by running it with mapping job configurations (i.e. source data configurations) and
-   * returns its results
-   *
-   * @param projectId project id the job belongs to
-   * @param jobId job id
-   * @param testResourceCreationRequest test resource creation request to be executed
-   * @return
-   */
-  def testMappingWithJob(projectId: String, jobId: String, testResourceCreationRequest: TestResourceCreationRequest): Future[Future[Seq[FhirMappingResult]]] = {
-    jobRepository.testMappingWithJob(projectId, jobId, testResourceCreationRequest)
-  }
-
-  /**
-   * Returns the logs of mapping tasks ran in the given execution.
-   *
-   * @param executionId the identifier of mapping job execution.
-   * @return the logs of mapping tasks
-   * */
-  def getExecutionLogs(executionId:String): Future[Seq[JValue]] ={
-    Future {
-      // read logs/tofhir-mappings.log file
-      val dataFrame = SparkConfig.sparkSession.read.json("logs/tofhir-mappings.log")
-      // handle the case where no job has been run yet which makes the data frame empty
-      if(dataFrame.isEmpty){
-        Seq.empty
-      }
-      else {
-        val jobRuns = dataFrame.filter(s"executionId = '$executionId'")
-        // handle the case where the job has not been run yet which makes the data frame empty
-        if (jobRuns.isEmpty) {
-          Seq.empty
-        } else {
-          jobRuns.collect().map(row => {
-            JsonMethods.parse(row.json)
-          })
-        }
-      }
-    }
-  }
-  /**
-   * Returns the list of mapping job executions. It extracts the logs from {@link logs/ tofhir - mappings.log} file for
-   * the given mapping job and groups them by their execution id and returns a single log for each execution. Further,
-   * it applies the pagination to the resulting execution logs.
-   *
-   * @param projectId project id the job belongs to
-   * @param jobId     job id
-   * @param queryParams parameters to filter results such as paging
-   * @return a tuple as follows
-   *         first element is the execution logs of mapping job as a JSON array. It returns an empty array if the job has not been run before.
-   *         second element is the total number of executions without applying any filters i.e. query params
-   * @throws ResourceNotFound when mapping job does not exist
-   */
-  def getExecutions(projectId: String, jobId: String, queryParams: Map[String, String]): Future[(Seq[JValue],Long)] = {
-    // retrieve the job to validate its existence
-    jobRepository.getJob(projectId, jobId).map {
-      case Some(_) =>
-        // read logs/tofhir-mappings.log file
-        val dataFrame = SparkConfig.sparkSession.read.json("logs/tofhir-mappings.log")
-        // handle the case where no job has been run yet which makes the data frame empty
-        if(dataFrame.isEmpty){
-          (Seq.empty,0)
-        }
-        else {
-          val jobRuns = dataFrame.filter(s"jobId = '$jobId' and projectId = '$projectId'")
-          // handle the case where the job has not been run yet which makes the data frame empty
-          if (jobRuns.isEmpty) {
-            (Seq.empty,0)
-          } else {
-            // group logs by execution id
-            val jobRunsGroupedByExecutionId = jobRuns.groupByKey(row => row.get(row.fieldIndex("executionId")).toString)(Encoders.STRING)
-            // get execution logs
-            val executionLogs = jobRunsGroupedByExecutionId.mapGroups((key,values) => {
-              // keeps the rows belonging to this execution
-              val rows = values.toList
-              val count = rows.length
-              val successCount = rows.count(r => r.get(r.fieldIndex("result")).toString.contentEquals("SUCCESS"))
-              // use the timestamp of first one, which is ran first, as timestamp of execution
-              val timestamp = rows.head.get(rows.head.fieldIndex("@timestamp")).toString
-              // set the status of execution
-              var status = "SUCCESS"
-              if(successCount == 0){
-                status = "FAILURE"
-              } else if(successCount != count){
-                status = "PARTIAL_SUCCESS"
-              }
-              Row.fromSeq(Seq(key,count,timestamp,status))
-            })(RowEncoder(StructType(
-              StructField("id",StringType) ::
-                StructField("mappingTaskCount",IntegerType) ::
-                StructField("timestamp",StringType) ::
-                StructField("status",StringType):: Nil
-            )))
-            // page size is 10, handle pagination
-            val total = executionLogs.count()
-            val numOfPages = Math.ceil(total.toDouble / 10).toInt
-            val page = queryParams.getOrElse("page", "1").toInt
-            // handle the case where requested page does not exist
-            if (page > numOfPages) {
-              (Seq.empty,0)
-            } else {
-              val start = (page - 1) * 10
-              val end = Math.min(start + 10, total.toInt)
-              // sort the executions by latest to oldest
-              val paginatedLogs = executionLogs.sort(executionLogs.col("timestamp").desc).collect().slice(start, end)
-              (paginatedLogs.map(row => {
-                JsonMethods.parse(row.json)
-              }),total)
-            }
-          }
-        }
-      case None => throw ResourceNotFound("Mapping job does not exists.", s"A mapping job with id $jobId does not exists")
-    }
   }
 }
